Add comments and 1024 LPT
